1：修改对象复制方法名； 2：增加拖拽修改排序的一些方法；

+++ warden-admin/src/main/java/com/microwarp/warden/stand/admin/domain/vo/SysDictionaryDataVO.java
@@ -20,7 +20,7 @@ public class SysDictionaryDataVO {
     /** 默认项 */
     private Boolean dataDefault;
     /** 排序 */
-    private Integer orders;
+    private Long orders;
     /** 是否禁用 */
     private Boolean disabled;
     /** 创建时间 */
@@ -84,11 +84,11 @@ public class SysDictionaryDataVO {
         this.dataDefault = dataDefault;
     }
 
-    public Integer getOrders() {
+    public Long getOrders() {
         return orders;
     }
 
-    public void setOrders(Integer orders) {
+    public void setOrders(Long orders) {
         this.orders = orders;
     }
