+++ warden-facade/src/main/java/com/microwarp/warden/stand/facade/syspost/dto/SysPostDTO.java
@@ -20,7 +20,7 @@ public class SysPostDTO implements Serializable {
     /** 描述 */
     private String description;
     /** 排序值 */
-    private Integer orders;
+    private Long orders;
     /** 创建时间 */
     private Date createDate;
     /** 修改时间 */
@@ -74,11 +74,11 @@ public class SysPostDTO implements Serializable {
         this.description = description;
     }
 
-    public Integer getOrders() {
+    public Long getOrders() {
         return orders;
     }
 
-    public void setOrders(Integer orders) {
+    public void setOrders(Long orders) {
         this.orders = orders;
     }
