+++ warden-service/src/main/java/com/microwarp/warden/stand/data/entity/SysDictionary.java
@@ -16,7 +16,7 @@ public class SysDictionary extends LogicEntity {
     /** 描述 */
     private String description;
     /** 排序值 */
-    private Integer orders;
+    private Long orders;
     /** 禁用 */
     private Boolean disabled;
 
@@ -44,11 +44,11 @@ public class SysDictionary extends LogicEntity {
         this.description = description;
     }
 
-    public Integer getOrders() {
+    public Long getOrders() {
         return orders;
     }
 
-    public void setOrders(Integer orders) {
+    public void setOrders(Long orders) {
         this.orders = orders;
     }
